image/image_metadata.py
# image_metadata.py
from PIL import Image  # Import PIL library for image processing

def get_image_metadata(image):
    """
    Placeholder function to get image metadata (e.g., height, speed).
    You need to implement this function to extract metadata from the image.
    For example, using libraries like PIL or ExifTool.
    This function should return a dictionary with metadata information.
    For simplicity, let's assume we return dummy data.
    """
    img = Image.open(image)
    height_meters = img.height * 0.0254  
    speed_meters_per_second = 6
    return {
        'height': height_meters,
        'speed': speed_meters_per_second
    }
