requirements.txt
asgiref==3.7.2
chardet==5.2.0
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
djangorestframework==3.14.0
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
